修改Notice拿取方式

+++ Fotricle/Models/ViewFeedBack.cs
@@ -15,13 +15,13 @@ namespace Fotricle.Models
         public int CustomerId { get; set; }
 
         [Display(Name = "餐點滿意程度")]
-        public int Food { set; get; }
+        public double Food { set; get; }
 
         [Display(Name = "服務滿意程度")]
-        public int Service { set; get; }
+        public double Service { set; get; }
 
         [Display(Name = "整體滿意程度")]
-        public int AllSuggest { set; get; }
+        public double AllSuggest { set; get; }
 
         [Display(Name = "餐車建議")]
         public string CarSuggest { set; get; }
